sort out the session logic a bit
