+++ app/controllers/application_controller.rb
@@ -1,8 +1,6 @@
 class ApplicationController < ActionController::API
   include AccessControl
-  include ActionController::Cookies
-
-  AUTHN_SESSION_NAME = 'authn'
+  include AuthNSession
 
   # Unauthenticated access means credentials are required but absent. It should map to the HTTP 401
   # status code.
@@ -22,30 +20,6 @@ class ApplicationController < ActionController::API
     URI.parse(request.referer).host
   rescue URI::InvalidURIError
     nil
-  end
-
-  private def establish_session(account_id, audience)
-    # avoid any potential session fixation. whatever session they had before can't be trusted.
-    RefreshToken.revoke(authn_session[:token]) if authn_session[:token]
-
-    # NOTE: this cookie is not set to expire, but the refresh token is.
-    cookies[AUTHN_SESSION_NAME] = {
-      value: JSON::JWT.new(
-        iss: Rails.application.config.authn_url,
-        sub: RefreshToken.create(account_id),
-        aud: Rails.application.config.authn_url,
-        iat: Time.now.utc.to_i,
-        azp: audience
-      ).sign(Rails.application.config.session_key, 'HS256').to_s,
-      secure: Rails.application.config.force_ssl,
-      httponly: true
-    }
-  end
-
-  private def authn_session
-    return {} unless cookies[AUTHN_SESSION_NAME].present?
-
-    @authn_session ||= JSON::JWT.decode(cookies[AUTHN_SESSION_NAME], Rails.application.config.session_key) || {}
   end
 
   private def issue_token_from(sess)
