+++ config/initializers/lib.rb
@@ -1,3 +1,4 @@
 require 'error_codes'
 require 'json_envelope'
 require 'access_control'
+require 'authn_session'
